Install typescript. Restructure project
