+++ client/src/components/App/App.tsx
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+import './App.css';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <header className="App-header">
+          <p>Welcome to Split!</p>
+        </header>
+      </div>
+    );
+  }
+}
+
+export default App;
